service/requirements.txt
Django>=3.2.9,<4.0
djangorestframework
django-cors-headers
requests
psycopg2-binary>=2.8
pytest
django-debug-toolbar
